+++ gluecodium/src/main/java/com/here/gluecodium/generator/dart/DartGeneratorSuite.kt
@@ -81,6 +81,7 @@ class DartGeneratorSuite(options: Gluecodium.Options) : GeneratorSuite {
         DartCommentsProcessor(options.werror.contains(Gluecodium.Options.WARNING_DOC_LINKS))
     private val overloadsWerror = options.werror.contains(Gluecodium.Options.WARNING_DART_OVERLOADS)
     private val testableMode = options.generateStubs
+    private val customTags = options.tags
 
     override fun generate(limeModel: LimeModel): List<GeneratedFile> {
         val limeLogger = LimeLogger(logger, limeModel.fileNameMap)
@@ -89,9 +90,14 @@ class DartGeneratorSuite(options: Gluecodium.Options) : GeneratorSuite {
         val ffiNameResolver = FfiNameResolver(limeModel.referenceMap, nameRules, internalPrefix)
 
         val ffiFilteredElements =
-            LimeModelFilter { it is LimeFunction || it is LimeProperty || !it.attributes.have(DART, SKIP) }
-                .filter(limeModel.topElements)
-        val dartFilteredElements = LimeModelFilter { !it.attributes.have(DART, SKIP) }.filter(limeModel.topElements)
+            LimeModelFilter {
+                !CommonGeneratorPredicates.hasSkipTags(it, customTags) &&
+                    (it is LimeFunction || it is LimeProperty || !it.attributes.have(DART, SKIP))
+            }.filter(limeModel.topElements)
+        val dartFilteredElements =
+            LimeModelFilter {
+                !CommonGeneratorPredicates.hasSkipTags(it, customTags) && !it.attributes.have(DART, SKIP)
+            }.filter(limeModel.topElements)
         val validationResult = DartOverloadsValidator(dartNameResolver, limeLogger, overloadsWerror)
             .validate(dartFilteredElements)
         if (!validationResult) {