+++ gluecodium/src/test/resources/smoke/skip/output/cbridge/include/smoke/cbridge_SkipTypesTags.h
@@ -1,0 +1,11 @@
+//
+//
+#pragma once
+#ifdef __cplusplus
+extern "C" {
+#endif
+#include "cbridge/include/BaseHandle.h"
+#include "cbridge/include/Export.h"
+#ifdef __cplusplus
+}
+#endif