+++ gluecodium/src/test/resources/smoke/skip/output/dart/ffi/ffi_smoke_SkipTagsOnly.h
@@ -1,0 +1,12 @@
+#pragma once
+#include "Export.h"
+#include "OpaqueHandle.h"
+#include <stdint.h>
+#ifdef __cplusplus
+extern "C" {
+#endif
+_GLUECODIUM_FFI_EXPORT FfiOpaqueHandle library_smoke_SkipTagsOnly_copy_handle(FfiOpaqueHandle handle);
+_GLUECODIUM_FFI_EXPORT void library_smoke_SkipTagsOnly_release_handle(FfiOpaqueHandle handle);
+#ifdef __cplusplus
+}
+#endif