+++ gluecodium/src/test/resources/smoke/skip/output/dart/ffi/ffi_smoke_SkipTypesTags.h
@@ -1,0 +1,10 @@
+#pragma once
+#include "Export.h"
+#include "OpaqueHandle.h"
+#include <stdint.h>
+#ifdef __cplusplus
+extern "C" {
+#endif
+#ifdef __cplusplus
+}
+#endif