+++ gluecodium/src/test/resources/smoke/skip/output/dart/lib/src/smoke/skip_types_tags.dart
@@ -1,0 +1,5 @@
+import 'dart:ffi';
+import 'package:ffi/ffi.dart';
+import 'package:meta/meta.dart';
+import 'package:library/src/_library_context.dart' as __lib;
+final bool placeHolder = true;